rename models to valueObjects
