+++ back/src/Domain/ValueObject/PostTitle.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Domain\Model;
+namespace App\Domain\ValueObject;
 
 use App\Domain\Exception\InvalidPostException;
